pythonTest0/eca-morphological/pyproject.toml
[project]
name = "eca-morphological"
version = "0.1.0"
description = ""
authors = [
    {name = "carlozrm90",email = "carloszrm90@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow (>=11.3.0,<12.0.0)",
    "numpy (>=2.2,<2.3.2)",
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "matplotlib (>=3.10.5,<4.0.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "jinja2 (>=3.1.6,<4.0.0)",
]
[tool.poetry]
packages = [{include = "eca_morphological", from = "src"}]

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.10"
pylint = "^3.3.8"
isort = "^6.0.1"
pre-commit = "^4.3.0"

[tool.ruff]
line-length = 100
target-version = "py313"
lint.extend-select = ["E","E4", "E7", "E9", "F", "B","PL"]
lint.fixable = ["ALL"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
